v3.1.0: Double→Long迁移 + 动画优化 + 预算超支修复
金额从Double分币迁移到Long整数分币，彻底消除浮点精度误差。
数据库MIGRATION_2_3自动转换旧数据。备份格式v2兼容v1旧版导入。
余额动画改用cubic ease-out曲线，启动更快。修复预算超支时始终显示¥0的问题。

+++ shared/src/commonMain/kotlin/com/example/expensetracker/data/ExpenseDao.kt
@@ -13,10 +13,10 @@ interface ExpenseDao {
     fun getAllExpenses(): Flow<List<Expense>>
 
     @Query("SELECT COALESCE(SUM(amount), 0) FROM expenses WHERE type = 'expense'")
-    fun getTotalExpense(): Flow<Double>
+    fun getTotalExpense(): Flow<Long>
 
     @Query("SELECT COALESCE(SUM(amount), 0) FROM expenses WHERE type = 'income'")
-    fun getTotalIncome(): Flow<Double>
+    fun getTotalIncome(): Flow<Long>
 
     @Insert
     suspend fun insert(expense: Expense)
